1. 위성 추적 데이터 콘솔 로그 및 파일로 저장되는 부분 완료. 2. 최대 각가속도, 각속도 추가완료. 현재 API와 연동 준비 중이므로 파일 저장되는 부분빼고 로그 남기는 부분 삭제 후 API 연계되도록 기능 개선 예정이므로 백업용도로 커밋함.

+++ src/main/kotlin/com/gtlsystems/acs_api/model/SatelliteTrackData.kt
@@ -28,8 +28,8 @@ object SatelliteTrackingData {
         val msInterval: Int
             get() = GlobalData.Tracking.msIntervel
 
-        val durationDays: Int
-            get() = GlobalData.Tracking.durationDays
+        val durationDays: Long
+            get() = GlobalData.Tracking.durationDays.toLong()
 
         val startDate: ZonedDateTime
             get() = GlobalData.Tracking.startDate
